+++ cmd/hottgo/main.go
@@ -6,7 +6,7 @@
 //	hottgo --check FILE        Type-check a file of S-expression terms
 //	hottgo --eval EXPR         Evaluate an S-expression term
 //	hottgo --synth EXPR        Synthesize the type of an S-expression term
-//	hottgo                     Start interactive REPL (TODO)
+//	hottgo                     Start interactive REPL
 package main
 
 import (
@@ -170,5 +170,9 @@ func repl() {
 			continue
 		}
 		fmt.Printf("%s\n", parser.FormatTerm(ty))
+	}
+
+	if err := scanner.Err(); err != nil {
+		fmt.Fprintf(os.Stderr, "error reading input: %v\n", err)
 	}
 }